feat: 分页重构 - 跨会话; 分页，发送，滑动分离

+++ app/src/main/java/info/hermiths/chatapp/model/req/HistoryBody.kt
@@ -11,6 +11,6 @@ data class HistoryBody(
 )
 
 data class SeqCondition(
-    val endSeq: Long,
-    val startSeq: Long,
+    val endSeq: Int,
+    val startSeq: Int,
 )
